+++ Cargo.toml
@@ -74,10 +74,9 @@ pgt_fs                     = { path = "./crates/pgt_fs", version = "0.0.0" }
 pgt_lexer                  = { path = "./crates/pgt_lexer", version = "0.0.0" }
 pgt_lexer_codegen          = { path = "./crates/pgt_lexer_codegen", version = "0.0.0" }
 pgt_lexer_new              = { path = "./crates/pgt_lexer_new", version = "0.0.0" }
+pgt_lexer_new_codegen      = { path = "./crates/pgt_lexer_new_codegen", version = "0.0.0" }
 pgt_lsp                    = { path = "./crates/pgt_lsp", version = "0.0.0" }
 pgt_markup                 = { path = "./crates/pgt_markup", version = "0.0.0" }
-pgt_parser                 = { path = "./crates/pgt_parser", version = "0.0.0" }
-pgt_parser_codegen         = { path = "./crates/pgt_parser_codegen", version = "0.0.0" }
 pgt_query_ext              = { path = "./crates/pgt_query_ext", version = "0.0.0" }
 pgt_query_ext_codegen      = { path = "./crates/pgt_query_ext_codegen", version = "0.0.0" }
 pgt_query_proto_parser     = { path = "./crates/pgt_query_proto_parser", version = "0.0.0" }
@@ -85,6 +84,7 @@ pgt_schema_cache           = { path = "./crates/pgt_schema_cache", version = "0.
 pgt_statement_splitter     = { path = "./crates/pgt_statement_splitter", version = "0.0.0" }
 pgt_text_edit              = { path = "./crates/pgt_text_edit", version = "0.0.0" }
 pgt_text_size              = { path = "./crates/pgt_text_size", version = "0.0.0" }
+pgt_tokenizer              = { path = "./crates/pgt_tokenizer", version = "0.0.0" }
 pgt_treesitter_queries     = { path = "./crates/pgt_treesitter_queries", version = "0.0.0" }
 pgt_typecheck              = { path = "./crates/pgt_typecheck", version = "0.0.0" }
 pgt_workspace              = { path = "./crates/pgt_workspace", version = "0.0.0" }
