+++ crates/pgt_lexer_new/src/codegen/syntax_kind.rs
@@ -1,0 +1,1 @@
+pgt_lexer_new_codegen::syntax_kind_codegen!();
