--- crates/pgt_parser/src/codegen/syntax_kind.rs
@@ -1,1 +1,0 @@
-pgt_parser_codegen::syntax_kind_codegen!();
