--- crates/pgt_parser/src/lib.rs
@@ -1,4 +1,0 @@
-mod codegen;
-mod lexed_str;
-
-pub use crate::codegen::syntax_kind::SyntaxKind;
