+++ crates/pgt_tokenizer/src/cursor.rs
